:alien:  btn salir

+++ src/app/login/login/login.component.ts
@@ -148,10 +148,12 @@ export class LoginComponent implements OnInit,OnDestroy {
 		this.eventService.sendApp(new EventMessage(1, new EventBlocked(type, msg)));
 	}
 	logout() {
-		localStorage.removeItem('user');
-		localStorage.clear();
-		this.router.navigate(['/']);
-		
-		dis.disenio();
+		if(this.appsIsActivate){
+			localStorage.removeItem('user');
+			localStorage.clear();
+			this.router.navigate(['/']);
+			
+			dis.disenio();
+		}
 	}
 }
